feature(@nestjs/core) expose container to ModuleRef, add close() to NestContext
