+++ packages/common/interfaces/nest-application-context.interface.ts
@@ -1,5 +1,4 @@
 import { Type } from './type.interface';
-import { LoggerService } from '../services/logger.service';
 
 export interface INestApplicationContext {
   /**
@@ -10,10 +9,17 @@ export interface INestApplicationContext {
 
   /**
    * Retrieves an instance of either injectable or controller available anywhere, otherwise, throws exception.
-   * @returns {T}
+   * @returns {TResult}
    */
-  get<T>(
-    typeOrToken: Type<T> | string | symbol,
+  get<TInput = any, TResult = TInput>(
+    typeOrToken: Type<TInput> | string | symbol,
     options?: { strict: boolean },
-  ): T;
+  ): TResult;
+
+  /**
+   * Terminates the application
+   *
+   * @returns {Promise<void>}
+   */
+  close(): Promise<void>;
 }
