+++ packages/core/injector/container.ts
@@ -2,9 +2,9 @@ import { DynamicModule } from '@nestjs/common';
 import { GLOBAL_MODULE_METADATA } from '@nestjs/common/constants';
 import { Type } from '@nestjs/common/interfaces/type.interface';
 import 'reflect-metadata';
+import { ApplicationConfig } from '../application-config';
+import { InvalidModuleException } from '../errors/exceptions/invalid-module.exception';
 import { UnknownModuleException } from '../errors/exceptions/unknown-module.exception';
-import { ApplicationConfig } from './../application-config';
-import { InvalidModuleException } from './../errors/exceptions/invalid-module.exception';
 import { ModuleCompiler } from './compiler';
 import { Module } from './module';
 import { ModulesContainer } from './modules-container';
