+++ packages/core/pipes/pipes-consumer.ts
@@ -1,6 +1,5 @@
-import { Transform, Paramtype } from '@nestjs/common/interfaces';
-import { RouteParamtypes } from '@nestjs/common/enums/route-paramtypes.enum';
-import { ParamsTokenFactory } from './../pipes/params-token-factory';
+import { Transform } from '@nestjs/common/interfaces';
+import { ParamsTokenFactory } from './params-token-factory';
 
 export class PipesConsumer {
   private readonly paramsTokenFactory = new ParamsTokenFactory();
