+++ packages/core/pipes/pipes-context-creator.ts
@@ -1,18 +1,10 @@
-import 'reflect-metadata';
-import iterate from 'iterare';
-import {
-  Controller,
-  PipeTransform,
-  Transform,
-} from '@nestjs/common/interfaces';
 import { PIPES_METADATA } from '@nestjs/common/constants';
-import {
-  isUndefined,
-  isFunction,
-  isEmpty,
-} from '@nestjs/common/utils/shared.utils';
-import { ApplicationConfig } from './../application-config';
-import { ContextCreator } from './../helpers/context-creator';
+import { Controller, PipeTransform, Transform } from '@nestjs/common/interfaces';
+import { isEmpty, isFunction, isUndefined } from '@nestjs/common/utils/shared.utils';
+import iterate from 'iterare';
+import 'reflect-metadata';
+import { ApplicationConfig } from '../application-config';
+import { ContextCreator } from '../helpers/context-creator';
 import { NestContainer } from '../injector/container';
 
 export class PipesContextCreator extends ContextCreator {
