+++ packages/core/router/router-exception-filters.ts
@@ -3,11 +3,11 @@ import { EXCEPTION_FILTERS_METADATA } from '@nestjs/common/constants';
 import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
 import { isEmpty } from '@nestjs/common/utils/shared.utils';
 import 'reflect-metadata';
+import { ApplicationConfig } from '../application-config';
 import { BaseExceptionFilterContext } from '../exceptions/base-exception-filter-context';
 import { ExceptionsHandler } from '../exceptions/exceptions-handler';
 import { NestContainer } from '../injector/container';
-import { ApplicationConfig } from './../application-config';
-import { RouterProxyCallback } from './../router/router-proxy';
+import { RouterProxyCallback } from './router-proxy';
 
 export class RouterExceptionFilters extends BaseExceptionFilterContext {
   constructor(
