+++ packages/core/router/router-explorer.ts
@@ -1,26 +1,26 @@
-import 'reflect-metadata';
-import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
+import { METHOD_METADATA, PATH_METADATA } from '@nestjs/common/constants';
 import { RequestMethod } from '@nestjs/common/enums/request-method.enum';
-import { RouterProxy, RouterProxyCallback } from './router-proxy';
-import { UnknownRequestMappingException } from '../errors/exceptions/unknown-request-mapping.exception';
+import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
 import { Type } from '@nestjs/common/interfaces/type.interface';
-import { isUndefined, validatePath } from '@nestjs/common/utils/shared.utils';
-import { RouterMethodFactory } from '../helpers/router-method-factory';
-import { PATH_METADATA, METHOD_METADATA } from '@nestjs/common/constants';
 import { Logger } from '@nestjs/common/services/logger.service';
+import { isUndefined, validatePath } from '@nestjs/common/utils/shared.utils';
+import 'reflect-metadata';
+import { ApplicationConfig } from '../application-config';
+import { UnknownRequestMappingException } from '../errors/exceptions/unknown-request-mapping.exception';
+import { GuardsConsumer } from '../guards/guards-consumer';
+import { GuardsContextCreator } from '../guards/guards-context-creator';
 import { routeMappedMessage } from '../helpers/messages';
-import { RouterExecutionContext } from './router-execution-context';
+import { RouterMethodFactory } from '../helpers/router-method-factory';
+import { NestContainer } from '../injector/container';
+import { InterceptorsConsumer } from '../interceptors/interceptors-consumer';
+import { InterceptorsContextCreator } from '../interceptors/interceptors-context-creator';
+import { MetadataScanner } from '../metadata-scanner';
+import { PipesConsumer } from '../pipes/pipes-consumer';
+import { PipesContextCreator } from '../pipes/pipes-context-creator';
 import { ExceptionsFilter } from './interfaces/exceptions-filter.interface';
 import { RouteParamsFactory } from './route-params-factory';
-import { MetadataScanner } from '../metadata-scanner';
-import { ApplicationConfig } from './../application-config';
-import { PipesContextCreator } from './../pipes/pipes-context-creator';
-import { PipesConsumer } from './../pipes/pipes-consumer';
-import { NestContainer } from '../injector/container';
-import { GuardsContextCreator } from '../guards/guards-context-creator';
-import { GuardsConsumer } from '../guards/guards-consumer';
-import { InterceptorsContextCreator } from '../interceptors/interceptors-context-creator';
-import { InterceptorsConsumer } from '../interceptors/interceptors-consumer';
+import { RouterExecutionContext } from './router-execution-context';
+import { RouterProxy, RouterProxyCallback } from './router-proxy';
 
 export class RouterExplorer {
   private readonly executionContextCreator: RouterExecutionContext;
