+++ packages/core/router/routes-resolver.ts
@@ -1,16 +1,16 @@
-import { NestContainer, InstanceWrapper } from '../injector/container';
-import { RouterProxy } from './router-proxy';
-import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
-import { Logger } from '@nestjs/common/services/logger.service';
-import { controllerMappingMessage } from '../helpers/messages';
-import { Resolver } from './interfaces/resolver.interface';
-import { RouterExceptionFilters } from './router-exception-filters';
-import { MetadataScanner } from '../metadata-scanner';
-import { RouterExplorer } from './router-explorer';
-import { ApplicationConfig } from './../application-config';
-import { NotFoundException, BadRequestException } from '@nestjs/common';
+import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { MODULE_PATH } from '@nestjs/common/constants';
 import { HttpServer } from '@nestjs/common/interfaces';
+import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
+import { Logger } from '@nestjs/common/services/logger.service';
+import { ApplicationConfig } from '../application-config';
+import { controllerMappingMessage } from '../helpers/messages';
+import { InstanceWrapper, NestContainer } from '../injector/container';
+import { MetadataScanner } from '../metadata-scanner';
+import { Resolver } from './interfaces/resolver.interface';
+import { RouterExceptionFilters } from './router-exception-filters';
+import { RouterExplorer } from './router-explorer';
+import { RouterProxy } from './router-proxy';
 
 export class RoutesResolver implements Resolver {
   private readonly logger = new Logger(RoutesResolver.name, true);
